+++ src/Modules/BabaPlay.Modules.MatchReports/DependencyInjection.cs
@@ -1,0 +1,13 @@
+using BabaPlay.Modules.MatchReports.Services;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace BabaPlay.Modules.MatchReports;
+
+public static class DependencyInjection
+{
+    public static IServiceCollection AddMatchReportsModule(this IServiceCollection services)
+    {
+        services.AddScoped<MatchReportService>();
+        return services;
+    }
+}